Minor fixes.

+++ restmote/sync.py
@@ -1,5 +1,4 @@
 import requests
-import time
 import logging
 import urlparse
 
@@ -7,11 +6,7 @@ from django.conf import settings
 
 logging.basicConfig(format='%(asctime)s %(levelname)s: %(message)s')
 
-f = open(settings.RESTMOTE_SNAP_FILE, 'r+')
 root = urlparse.urljoin(settings.RESTMOTE_HOST + ":" + settings.RESTMOTE_PORT, settings.RESTMOTE_API_ROOT)
-get_params = "?" + settings.RESTMOTE_FILTER_FIELD + "=" + f.read().strip()
-
-f.close()
 
 
 def get_data(url):
@@ -19,19 +14,17 @@ def get_data(url):
     if r.ok:
         logging.info(url)
         logging.info(r.json())
-        return r.json()
+        return True, r.json()
     else:
-        return []
+        logging.info("Connection failed: %s" % r.text)
+        return False, []
 
 
-def build_objects(obj_class, obj_string, data, field_bindings, nested=[], key=None):
+def build_objects(obj_class, obj_string, data, field_bindings, nested=[]):
     for e in data:
 
         try:
-            if key:
-                o = obj_class.objects.get(**{'id' + obj_string: e[key]["id"]})
-            else:
-                o = obj_class.objects.get(**{'id' + obj_string: e["id"]})
+            o = obj_class.objects.get(**{'id' + obj_string: e["id"]})
         except obj_class.DoesNotExist:
             o = obj_class()
 
@@ -42,26 +35,28 @@ def build_objects(obj_class, obj_string, data, field_bindings, nested=[], key=No
             for f in [x for x in e[n] if x in field_bindings]:
                 setattr(o, field_bindings[f], e[n][f])
 
-        if key:
-            setattr(o, "id" + obj_string, e[key]["id"])
-        else:
-            setattr(o, "id" + obj_string, e["id"])
+        setattr(o, "id" + obj_string, e["id"])
         o.save()
+        logging.info("Added %s: %s" % (obj_string, o.pk))
 
 
-def sync_objects(url, obj_class, obj_string, field_bindings, nested=[], key=None):
-    data = get_data(root + url + get_params)
-    build_objects(obj_class, obj_string, data, field_bindings, nested, key)
+def sync_objects(url, qfilter, obj_class, obj_string, field_bindings, nested=[]):
+    status, data = get_data(root + url + '?' + qfilter)
+    if status:
+        build_objects(obj_class, obj_string, data, field_bindings, nested)
+        return 1
+    else:
+        return 0
 
 
 def remove_objects(url, obj_class, obj_string):
-    remote_ids = get_data(root + url)
-    local_ids = obj_class.objects.values_list('id' + obj_string, flat=True)
-    must_remove = list(set(local_ids).difference(remote_ids))
-    obj_class.objects.filter(**{'id' + obj_string + '__in': must_remove}).delete()
-
-# Order IS important!! Care with relationships
-
-f = open(settings.RESTMOTE_SNAP_FILE, 'w')
-f.write(time.strftime("%Y-%m-%d %H:%M", (time.localtime(time.time()))))
-f.close()
+    status, remote_ids = get_data(root + url)
+    if status:
+        local_ids = obj_class.objects.values_list('id' + obj_string, flat=True)
+        must_remove = list(set(local_ids).difference(remote_ids))
+        obj_class.objects.filter(**{'id' + obj_string + '__in': must_remove}).delete()
+        if must_remove:
+            logging.info("Deleted %s: %s" % (obj_string, ', '.join(must_remove)))
+        return 1
+    else:
+        return 0
